Make expo-web-browser easier to unit test.
