+++ packages/expo-web-browser/android/build.gradle
@@ -1,12 +1,23 @@
 apply plugin: 'com.android.library'
 apply plugin: 'maven'
+apply plugin: 'kotlin-android'
 
 group = 'host.exp.exponent'
 version = '8.2.0'
 
-// Simple helper that allows the root project to override versions declared by this library.
-def safeExtGet(prop, fallback) {
-  rootProject.ext.has(prop) ? rootProject.ext.get(prop) : fallback
+buildscript {
+  // Simple helper that allows the root project to override versions declared by this library.
+  ext.safeExtGet = { prop, fallback ->
+    rootProject.ext.has(prop) ? rootProject.ext.get(prop) : fallback
+  }
+
+  repositories {
+    mavenCentral()
+  }
+
+  dependencies {
+    classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:${safeExtGet("kotlinVersion", "1.3.50")}")
+  }
 }
 
 // Upload android library to maven with javadoc and android sources
@@ -50,6 +61,11 @@ android {
     sourceCompatibility = '1.8'
     targetCompatibility = '1.8'
   }
+
+  useLibrary 'android.test.runner'
+
+  useLibrary 'android.test.base'
+  useLibrary 'android.test.mock'
 }
 
 if (new File(rootProject.projectDir.parentFile, 'package.json').exists()) {
@@ -63,5 +79,14 @@ if (new File(rootProject.projectDir.parentFile, 'package.json').exists()) {
 dependencies {
   unimodule 'unimodules-core'
 
-  api "androidx.browser:browser:1.0.0"
+  // Core library
+  // Required -- JUnit 4 framework
+  testImplementation 'junit:junit:4.12'
+  // Optional -- Robolectric environment
+//  testImplementation 'androidx.test:core:1.2.0'
+  // Optional -- Mockito framework
+  testImplementation project(':unimodules-test-core')
+
+  api "androidx.browser:browser:1.2.0"
+  api "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${safeExtGet("kotlinVersion", "1.3.50")}"
 }
